chore: move Advanced Search specific functions to AdvancedSearchHelper
